// src/interfaces/models/luong.model.ts
export interface ILuongModel {
    maNhanVien: number;
    thang: number;
    nam: number;
    soNgayLe?: number;
    soNgayNghi?: number;
    soNgayCong: number;
    congChuanCuaThang: number;
    soGioTangCa?: number;
    tienThuong?: number;
    tongTienPhuCap?: number;
    tienTru?: number;
    tongTienTangCa?: number;
    luongGio?: number;
    soNgayNghiCoPhep?: number,
    luongTheoNgay: number;
    luongNgayNghi: number;
    soNgayNghiCoLuong: number
    soNgayNghiTruLuong: number;
    tongLuong: number;
    tongSoTienNgayNghiTruLuong: number;
}